Add policy

+++ app/Http/Controllers/TagsController.php
@@ -14,11 +14,14 @@ namespace App\Http\Controllers;
 
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagsController extends Controller
 {
     public function index(Request $request)
     {
+        Gate::authorize('viewAny', Tag::class);
+
         $query = Tag::query();
 
         $q = $request->query('q');
@@ -44,6 +47,8 @@ class TagsController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('create', Tag::class);
+
         $request->validate([
             'name' => 'required',
         ]);
@@ -60,6 +65,8 @@ class TagsController extends Controller
 
     public function show(Request $request, Tag $tag)
     {
+        Gate::authorize('view', $tag);
+
         return view('pages.tags-show', [
             'tag' => $tag,
         ]);
@@ -67,6 +74,8 @@ class TagsController extends Controller
 
     public function edit(Request $request, Tag $tag)
     {
+        Gate::authorize('update', $tag);
+
         return view('pages.tags-edit', [
             'tag' => $tag,
         ]);
@@ -74,6 +83,8 @@ class TagsController extends Controller
 
     public function update(Request $request, Tag $tag)
     {
+        Gate::authorize('update', $tag);
+
         $request->validate([
             'name' => 'required|min:2',
         ]);
@@ -89,6 +100,8 @@ class TagsController extends Controller
 
     public function destroy(Request $request, Tag $tag)
     {
+        Gate::authorize('delete', $tag);
+
         $tag->delete();
 
         return redirect()->back()->with('success', __('record_deleted_message'));
